Initial commit for ETH Global Taipei hackathon. The code in this commit was auto-generated by Bolt dot new with the following prompt:
"""
a simple web app in next to clone the robinhood experience to buy tokens on chain, displaying information of the token and supporting cross chain swaps without the user knowing.
Only write the frontend and mock all calls to the blockchain. Assume a user is connected with their wallet and funded.

Make the landing clean, with a simple list of assets, pnl per assets and aggrete.

When user clicks an asset in the list, a page opens with the profile of the token with information, graph and list of transactions.

Include a buy token tab where the user is prompted a list of tokens to buy.

Make it look like startup in San Francisco would.

use nextjs.
""":